InventoryService/python/requirements.txt
flask==2.1.1
pandas==1.4.2
sklearn==0.24.2
mysql-connector-python==8.0.28
